+++ src/modules/config/__tests__/config.service.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it } from 'vitest';
 
 import { ConfigService } from '../config.service.js';
 
-import type { RuntimeEnv } from '../../../schemas/env.schema.js';
+import type { RuntimeEnv } from '../../../user-types.js';
 
 describe('ConfigService', () => {
   let configService: ConfigService;
